+++ NEWS.md
@@ -1,4 +1,4 @@
-# tidycreel (development version)
+# tidycreel 5.2.0 "River Carpsucker" (2026-08-28)
 
 ## Breaking changes